Don't start a shell script every boot cycle.
This change makes the maria DB init script start from a
separate unit that is conditional. This way, when the
system starts mariadb at first, the script is run to
setup the DB, but on subsequent startups, the shell
script isn't started. Removal of the entire DB will
trigger the script to be run again automatically.